RC: refine All Actions test to be more precise
